+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.5
+
+* fix restoration indices
+
+## 0.0.4
+
+* save/restore pane state 
+
+## 0.0.3
+
+* Fix dynamic pane position and visibility
+
 ## 0.0.2
 
 * Updated dart docs